Notes on Keys
Handling List Updates:

Option A: 

Remove all existing HTML tied to this list. 

Recreate all HTML elements.

Option B: Apply a "Key" to each element during the mapping step.

After re-rendering, compare the keys on each ImageShow to the keys from the previous render. 

Apply minimal set of changes to existing DOM elements.

Requirements for Keys

Use whenever we have a of elements (so every time we do a 'map').

Add key to the top-most JSX element in the list.

Must be a string or number.

Should be unique for this list.

Should be consistent across rerenders.